Thu, Jan 6, 2022, 5:27 PM +00:00

+++ readability/readability.c
@@ -16,12 +16,12 @@ int main(void)
     printf("%i sentences\n", count_sentences(text));
 
 
-    float per_hundred = 100 / count_words(text); //7.14
-    printf("%f\n", per_hundred);
-    float L = round(count_letters(text) / count_words(text)) * per_hundred;    //average number of letters per 100 words
-    printf("%f\n", L);
-    float S = round(count_sentences(text) / count_words(text)) * per_hundred;  //average number of sentences per 100 words
-    printf("%f\n", S);
+    int per_hundred = 100 / count_words(text); //7.14
+    printf("%i\n", per_hundred);
+    int L = round(count_letters(text) / count_words(text)) * per_hundred;    //average number of letters per 100 words
+    printf("%i\n", L);
+    int S = round(count_sentences(text) / count_words(text)) * per_hundred;  //average number of sentences per 100 words
+    printf("%i\n", S);
     int index = 0.0588 * L - 0.296 * S - 15.8;
 
     printf("Grade %i\n", index);
